1차 기본 기능 & 모델 기능 병합 #7
