+++ main/member.py
@@ -29,20 +29,20 @@ blueprint = Blueprint("member", __name__, url_prefix='')
 ########################################################################
 
 
-def authorize(f):
-    @wraps(f)
-    def decorated_function():
-        if not 'Authorization' in request.headers:  # headers 에서 Authorization 인증을 하고
-            abort(401)  # Authorization 으로 토큰이 오지 않았다면 에러 401
-        # Authorization 이 headers에 있다면 token 값을 꺼내온다.
-        token = request.headers['Authorization']
-        try:
-            user = jwt.decode(token, SECRET_KEY, algorithms=[
-                              'HS256'])  # 꺼내온 token 값을 디코딩해서 꺼내주고
-        except:
-            abort(401)  # 디코딩이 안될 경우 에러 401
-        return f(user)
-    return decorated_function
+# def authorize(f):
+#     @wraps(f)
+#     def decorated_function():
+#         if not 'Authorization' in request.headers:  # headers 에서 Authorization 인증을 하고
+#             abort(401)  # Authorization 으로 토큰이 오지 않았다면 에러 401
+#         # Authorization 이 headers에 있다면 token 값을 꺼내온다.
+#         token = request.headers['Authorization']
+#         try:
+#             user = jwt.decode(token, SECRET_KEY, algorithms=[
+#                               'HS256'])  # 꺼내온 token 값을 디코딩해서 꺼내주고
+#         except:
+#             abort(401)  # 디코딩이 안될 경우 에러 401
+#         return f(user)
+#     return decorated_function
 
 
 ########################################################################
@@ -150,7 +150,7 @@ def login():
     # 토큰 정의
     payload = {
         'id': str(result['_id']),
-        'exp': datetime.utcnow() + timedelta(seconds=60 * 60 * 24)  # 토큰 시간 적용
+        'exp': datetime.datetime.utcnow() + timedelta(seconds=60 * 60 * 24)  # 토큰 시간 적용
     }
     token = jwt.encode(payload, SECRET_KEY, algorithm='HS256')
     # print(token)
@@ -166,7 +166,7 @@ def login():
 ########################################################################
 ########################################################################
 @blueprint.route("/getuserinfo", methods=["GET"])
-@authorize
+@config.authorize
 def get_user_info(user):
     result = db.member.find_one({
         '_id': ObjectId(user['id'])
